final commit

+++ ft_strlcpy.c
@@ -6,16 +6,15 @@
 /*   By: prigaudi <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/05 13:55:43 by prigaudi          #+#    #+#             */
-/*   Updated: 2024/11/15 14:41:33 by prigaudi         ###   ########.fr       */
+/*   Updated: 2024/11/20 18:08:08 by prigaudi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
-#include <stdio.h>
 
 size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
-	unsigned long	i;
+	size_t	i;
 
 	if (size == 0)
 		return (ft_strlen(src));
